requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
python-dotenv==0.20.0
pycountry==22.3.5
requests==2.27.1